fix: hide click context menu ico in col-header in readonly mode
