+++ packages/core/src/utils/index.ts
@@ -235,3 +235,26 @@ export function chatatABC(n: number) {
 
   return s.toUpperCase();
 }
+
+export function isAllowEdit(
+  ctx: Context,
+  range?: Sheet["luckysheet_select_save"]
+) {
+  const cfg = ctx.config;
+  const judgeRange = _.isUndefined(range) ? ctx.luckysheet_select_save : range;
+  return (
+    _.every(judgeRange, (selection) => {
+      for (let r = selection.row[0]; r <= selection.row[1]; r += 1) {
+        if (cfg.rowReadOnly?.[r]) {
+          return false;
+        }
+      }
+      for (let c = selection.column[0]; c <= selection.column[1]; c += 1) {
+        if (cfg.colReadOnly?.[c]) {
+          return false;
+        }
+      }
+      return true;
+    }) && (_.isUndefined(ctx.allowEdit) ? true : ctx.allowEdit)
+  );
+}
